requirements.txt
requests
pytrends
markdown
jinja2
pyyaml
python-frontmatter
